--- Crojecter/src/com/kh/board/gallery/model/GalleryDao.java
@@ -1,5 +1,0 @@
-package com.kh.board.gallery.model;
-
-public class GalleryDao {
-
-}
