+++ Crojecter/src/com/kh/board/gallery/model/dao/GalleryDao.java
@@ -1,0 +1,5 @@
+package com.kh.board.gallery.model.dao;
+
+public class GalleryDao {
+
+}
