+++ Crojecter/src/com/kh/board/gallery/model/service/GalleryService.java
@@ -1,0 +1,5 @@
+package com.kh.board.gallery.model.service;
+
+public class GalleryService {
+
+}
