--- Crojecter/src/com/kh/board/notice/model/NoticeDao.java
@@ -1,5 +1,0 @@
-package com.kh.board.notice.model;
-
-public class NoticeDao {
-
-}
