--- Crojecter/src/com/kh/board/notice/model/NoticeService.java
@@ -1,5 +1,0 @@
-package com.kh.board.notice.model;
-
-public class NoticeService {
-
-}
