+++ Crojecter/src/com/kh/board/notice/model/dao/NoticeDao.java
@@ -1,0 +1,5 @@
+package com.kh.board.notice.model.dao;
+
+public class NoticeDao {
+
+}
