+++ Crojecter/src/com/kh/board/notice/model/service/NoticeService.java
@@ -1,0 +1,5 @@
+package com.kh.board.notice.model.service;
+
+public class NoticeService {
+
+}
