--- Crojecter/src/com/kh/board/project/model/ProjectDao.java
@@ -1,5 +1,0 @@
-package com.kh.board.project.model;
-
-public class ProjectDao {
-
-}
