+++ Crojecter/src/com/kh/board/project/model/dao/ProjectDao.java
@@ -1,0 +1,5 @@
+package com.kh.board.project.model.dao;
+
+public class ProjectDao {
+
+}
