+++ Crojecter/src/com/kh/board/project/model/service/ProjectService.java
@@ -1,0 +1,5 @@
+package com.kh.board.project.model.service;
+
+public class ProjectService {
+
+}
